+++ src/main/java/com/oracle/oracle/training/repository/UserProfileRepository.java
@@ -3,6 +3,6 @@ package com.oracle.oracle.training.repository;
 import com.oracle.oracle.training.entity.UserProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserProfileRepository extends JpaRepository<UserProfile,Long> {
-    boolean existsById(Long id);
+public interface UserProfileRepository extends JpaRepository<UserProfile,Integer> {
+    boolean existsById(Integer id);
 }
